requirements.txt
ensure==1.0.4
fastapi==0.110.1
imagehash==4.3.1
opencv_python==4.9.0.80
pandas==2.2.1
Pillow==10.3.0
pygame==2.5.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.32.2
ultralytics==8.1.42
uvicorn==0.29.0
